requirements/default.txt
flask
flask-classful
flask-sqlalchemy
flask-migrate
flask-cors
flask-request-params
flask-marshmallow
gunicorn
psycopg2-binary
python-dotenv
pyjwt
